# openapi.yaml
openapi: 3.0.3
info:
  title: Point of Sale (POS) System API
  version: 1.0.0
  description: >
    Comprehensive REST API documentation for POS System Backend featuring multi-branch management,
    real-time analytics, inventory control, cashier shift management, refund processing, and payment integration.
  contact:
    name: POS System Support
    email: support@pos-system.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
security:
  - BearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: User Signup
      description: Registers a new user (Store Admin, Branch Manager, Cashier, etc.) in the POS system and returns a JWT authentication token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '201':
          description: User successfully registered and JWT returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad Request / Email already exists
  /auth/signin:
    post:
      tags:
        - Authentication
      summary: User Signin / Login
      description: Authenticates user credentials (email & password) and returns a valid JWT authentication token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: Authentication successful and JWT returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials / Unauthorized

  # ---------------------------------------------------------------------------
  # User Profile Management
  # ---------------------------------------------------------------------------
...
  /api/branch-analytics/daily-sales:
    get:
      tags:
        - Branch Analytics
      summary: Get Daily Sales Chart
      description: Retrieves daily sales trends for a branch over a specified number of days (default 7).
      parameters:
        - name: branchId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Successfully retrieved daily sales chart data
  /api/branch-analytics/top-products:
    get:
      tags:
        - Branch Analytics
      summary: Get Top Products by Quantity
      description: Retrieves the top-selling products by quantity along with percentage contribution for a specific branch.
      parameters:
        - name: branchId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved top products performance data
  /api/branch-analytics/top-cashiers:
    get:
      tags:
        - Branch Analytics
      summary: Get Top Cashiers Performance
      description: Retrieves cashier rankings and performance metrics based on processed order count for a branch.
      parameters:
        - name: branchId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved top cashiers performance data
  /api/branch-analytics/category-sales:
    get:
      tags:
        - Branch Analytics
      summary: Get Category-wise Sales Breakdown
      description: Retrieves product category sales breakdown for a branch on a given date.
      parameters:
        - name: branchId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successfully retrieved category sales breakdown
  /api/branch-analytics/today-overview:
    get:
      tags:
        - Branch Analytics
      summary: Get Today's Branch Overview
      description: Retrieves today's dashboard overview metrics (total sales, order count, active shift count, etc.) for a branch.
      parameters:
        - name: branchId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved today's branch overview
  /api/branch-analytics/payment-breakdown:
    get:
      tags:
        - Branch Analytics
      summary: Get Payment Method Breakdown
      description: Retrieves payment method breakdown (Cash, Card, UPI, etc.) for a branch on a specific date.
      parameters:
        - name: branchId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successfully retrieved payment method breakdown

  # ---------------------------------------------------------------------------
  # Store Analytics
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT Bearer token obtained from `/auth/login` or `/auth/signup`
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [ROLE_ADMIN, ROLE_STORE_ADMIN, ROLE_STORE_MANAGER, ROLE_BRANCH_MANAGER, ROLE_BRANCH_CASHIER]
        storeId:
          type: integer
          format: int64
        branchId:
          type: integer
          format: int64
    AuthResponse:
      type: object
      properties:
        jwt:
          type: string
        message:
          type: string
        role:
          type: string
...